Rename proxy to server for clarity
