+++ server/valid.go
@@ -1,4 +1,4 @@
-package proxy
+package server
 
 import (
 	"github.com/pkg/errors"
